# openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Mars Vista API",
    "description": "Unified API for NASA Mars rover photos with advanced filtering, location queries, and scientific metadata. Access 680,000+ photos from Curiosity, Perseverance, Opportunity, and Spirit rovers.",
    "contact": {
      "name": "Mars Vista API",
      "url": "https://marsvista.dev"
    },
    "version": "v2.0.0"
  },
  "paths": {
...
    "/api/v2/panoramas": {
      "get": {
        "tags": [
          "V2 - Advanced Features"
        ],
        "summary": "Get all detected panoramas",
        "parameters": [
          {
            "name": "rovers",
            "in": "query",
            "description": "Comma-separated list of rover names (curiosity, perseverance, etc.)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sol_min",
            "in": "query",
            "description": "Minimum sol",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "sol_max",
            "in": "query",
            "description": "Maximum sol",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "min_photos",
            "in": "query",
            "description": "Only return panoramas with at least this many photos",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "page",
            "in": "query",
            "description": "Page number (1-indexed)",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 1
            }
          },
          {
            "name": "per_page",
            "in": "query",
            "description": "Items per page (default: 25, max: 100)",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 25
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/PanoramaResourceListApiResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PanoramaResourceListApiResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/PanoramaResourceListApiResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/panoramas/{id}": {
      "get": {
        "tags": [
          "V2 - Advanced Features"
        ],
        "summary": "Get a specific panorama by ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Panorama ID (e.g., pano_curiosity_1000_14)",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/PanoramaResource"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PanoramaResource"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/PanoramaResource"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ApiError": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "description": "A URI reference that identifies the problem type",
            "nullable": true
          },
          "title": {
            "type": "string",
            "description": "A short, human-readable summary of the problem type",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "description": "The HTTP status code",
            "format": "int32"
          },
          "detail": {
            "type": "string",
            "description": "A human-readable explanation specific to this occurrence of the problem",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "description": "A URI reference that identifies the specific occurrence of the problem",
            "nullable": true
          },
          "errors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "description": "Field-level validation errors",
            "nullable": true
          },
          "timestamp": {
            "type": "string",
            "description": "Response timestamp",
            "format": "date-time"
          }
        },
        "additionalProperties": false,
        "description": "RFC 7807 Problem Details for HTTP APIs\r\nProvides machine-readable error responses with detailed field-level validation errors"
      },
...
      "CameraAttributes": {
        "type": "object",
        "properties": {
          "full_name": {
            "type": "string",
            "description": "Full camera name",
            "nullable": true
          },
          "photo_count": {
            "type": "integer",
            "description": "Number of photos from this camera (when included in statistics)",
            "format": "int32",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "Camera attributes"
      },
      "CameraReference": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Camera identifier (name like \"FHAZ\", \"MAST\")",
            "nullable": true
          },
          "type": {
            "type": "string",
            "description": "Resource type (always \"camera\")",
            "nullable": true
          },
          "attributes": {
            "$ref": "#/components/schemas/CameraAttributes"
          }
        },
        "additionalProperties": false,
        "description": "Camera reference with attributes"
      },
...
      "CursorInfo": {
        "type": "object",
        "properties": {
          "current": {
            "type": "string",
            "description": "Cursor for the current page",
            "nullable": true
          },
          "next": {
            "type": "string",
            "description": "Cursor for the next page",
            "nullable": true
          },
          "previous": {
            "type": "string",
            "description": "Cursor for the previous page",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "Cursor-based pagination pointers"
      },
...
      "PaginationInfo": {
        "type": "object",
        "properties": {
          "page": {
            "type": "integer",
            "description": "Current page number (1-indexed)",
            "format": "int32",
            "nullable": true
          },
          "per_page": {
            "type": "integer",
            "description": "Number of items per page",
            "format": "int32"
          },
          "total_pages": {
            "type": "integer",
            "description": "Total number of pages",
            "format": "int32",
            "nullable": true
          },
          "cursor": {
            "$ref": "#/components/schemas/CursorInfo"
          }
        },
        "additionalProperties": false,
        "description": "Pagination information supporting both page-based and cursor-based pagination"
      },
      "PanoramaAttributes": {
        "type": "object",
        "properties": {
          "rover": {
            "type": "string",
            "description": "Rover name",
            "nullable": true
          },
          "sol": {
            "type": "integer",
            "description": "Sol when panorama was taken",
            "format": "int32"
          },
          "mars_time_start": {
            "type": "string",
            "description": "Mars time when panorama sequence started",
            "nullable": true
          },
          "mars_time_end": {
            "type": "string",
            "description": "Mars time when panorama sequence ended",
            "nullable": true
          },
          "total_photos": {
            "type": "integer",
            "description": "Total number of photos in the panorama",
            "format": "int32"
          },
          "coverage_degrees": {
            "type": "number",
            "description": "Approximate angular coverage in degrees",
            "format": "float",
            "nullable": true
          },
          "location": {
            "$ref": "#/components/schemas/PhotoLocation"
          },
          "camera": {
            "type": "string",
            "description": "Camera used for the panorama",
            "nullable": true
          },
          "avg_elevation": {
            "type": "number",
            "description": "Average elevation angle",
            "format": "float",
            "nullable": true
          },
          "unique_positions": {
            "type": "integer",
            "description": "Number of unique azimuth positions (distinct camera angles)",
            "format": "int32"
          },
          "avg_position_spacing": {
            "type": "number",
            "description": "Average spacing between positions in degrees",
            "format": "float",
            "nullable": true
          },
          "quality": {
            "type": "string",
            "description": "Quality tier based on coverage and positions: partial, half, wide, or full",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "Panorama attributes"
      },
      "PanoramaLinks": {
        "type": "object",
        "properties": {
          "stitched_preview": {
            "type": "string",
            "description": "Link to panorama preview/stitched image (future feature)",
            "nullable": true
          },
          "download_set": {
            "type": "string",
            "description": "Link to download all photos as a set",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "Links related to the panorama"
      },
      "PanoramaResource": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Unique panorama identifier (e.g., \"pano_curiosity_1000_14\")",
            "nullable": true
          },
          "type": {
            "type": "string",
            "description": "Resource type (always \"panorama\")",
            "nullable": true
          },
          "attributes": {
            "$ref": "#/components/schemas/PanoramaAttributes"
          },
          "photos": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PhotoResource"
            },
            "description": "Photos that make up this panorama",
            "nullable": true
          },
          "links": {
            "$ref": "#/components/schemas/PanoramaLinks"
          }
        },
        "additionalProperties": false,
        "description": "Panorama resource representing an auto-detected panoramic sequence"
      },
      "PanoramaResourceListApiResponse": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PanoramaResource"
            },
            "description": "The primary data for this response",
            "nullable": true
          },
          "meta": {
            "$ref": "#/components/schemas/ResponseMeta"
          },
          "pagination": {
            "$ref": "#/components/schemas/PaginationInfo"
          },
          "links": {
            "$ref": "#/components/schemas/ResponseLinks"
          }
        },
        "additionalProperties": false,
        "description": "Standardized API response envelope for v2 endpoints\r\nInspired by JSON:API specification for consistency"
      },
...
      "PhotoAttributes": {
        "type": "object",
        "properties": {
          "nasa_id": {
            "type": "string",
            "description": "NASA's unique identifier for this photo",
            "nullable": true
          },
          "sol": {
            "type": "integer",
            "description": "Mars sol (day) when photo was taken",
            "format": "int32",
            "nullable": true
          },
          "earth_date": {
            "type": "string",
            "description": "Earth date when photo was taken (YYYY-MM-DD)",
            "nullable": true
          },
          "date_taken_utc": {
            "type": "string",
            "description": "UTC timestamp when photo was taken",
            "format": "date-time",
            "nullable": true
          },
          "date_taken_mars": {
            "type": "string",
            "description": "Mars local time when photo was taken (e.g., \"Sol-1000M14:23:45\")",
            "nullable": true
          },
          "images": {
            "$ref": "#/components/schemas/PhotoImages"
          },
          "dimensions": {
            "$ref": "#/components/schemas/PhotoDimensions"
          },
          "sample_type": {
            "type": "string",
            "description": "Sample type (e.g., \"Full\", \"Thumbnail\", \"Subframe\")",
            "nullable": true
          },
          "location": {
            "$ref": "#/components/schemas/PhotoLocation"
          },
          "telemetry": {
            "$ref": "#/components/schemas/PhotoTelemetry"
          },
          "title": {
            "type": "string",
            "description": "Photo title",
            "nullable": true
          },
          "caption": {
            "type": "string",
            "description": "Photo caption/description",
            "nullable": true
          },
          "credit": {
            "type": "string",
            "description": "Photo credit (e.g., \"NASA/JPL-Caltech\")",
            "nullable": true
          },
          "created_at": {
            "type": "string",
            "description": "When this photo was added to our database",
            "format": "date-time",
            "nullable": true
          },
          "img_src": {
            "type": "string",
            "description": "URL to the photo image (legacy field, use images.medium instead)",
            "nullable": true
          },
          "raw_data": {
            "description": "Raw NASA API response data (JSONB). Only included when field_set=complete.\r\nContains all original NASA fields for data validation and debugging.",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "Photo attributes (data fields)"
      },
      "PhotoCoordinates": {
        "type": "object",
        "properties": {
          "x": {
            "type": "number",
            "description": "X coordinate",
            "format": "float"
          },
          "y": {
            "type": "number",
            "description": "Y coordinate",
            "format": "float"
          },
          "z": {
            "type": "number",
            "description": "Z coordinate",
            "format": "float"
          }
        },
        "additionalProperties": false,
        "description": "3D coordinates"
      },
      "PhotoDimensions": {
        "type": "object",
        "properties": {
          "width": {
            "type": "integer",
            "description": "Image width in pixels",
            "format": "int32"
          },
          "height": {
            "type": "integer",
            "description": "Image height in pixels",
            "format": "int32"
          }
        },
        "additionalProperties": false,
        "description": "Image dimensions"
      },
      "PhotoImages": {
        "type": "object",
        "properties": {
          "small": {
            "type": "string",
            "description": "Small size (320px wide) - for thumbnails",
            "nullable": true
          },
          "medium": {
            "type": "string",
            "description": "Medium size (800px wide) - for galleries",
            "nullable": true
          },
          "large": {
            "type": "string",
            "description": "Large size (1200px wide) - for detailed viewing",
            "nullable": true
          },
          "full": {
            "type": "string",
            "description": "Full resolution - for download and analysis",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "Multiple image URLs for different sizes"
      },
      "PhotoLocation": {
        "type": "object",
        "properties": {
          "site": {
            "type": "integer",
            "description": "Site number (geological location marker)",
            "format": "int32",
            "nullable": true
          },
          "drive": {
            "type": "integer",
            "description": "Drive number (rover's drive sequence)",
            "format": "int32",
            "nullable": true
          },
          "coordinates": {
            "$ref": "#/components/schemas/PhotoCoordinates"
          }
        },
        "additionalProperties": false,
        "description": "Location where photo was taken"
      },
      "PhotoMeta": {
        "type": "object",
        "properties": {
          "is_panorama_part": {
            "type": "boolean",
            "description": "Whether this photo is part of a panorama sequence",
            "nullable": true
          },
          "panorama_sequence_id": {
            "type": "string",
            "description": "Panorama sequence identifier if part of a panorama",
            "nullable": true
          },
          "has_stereo_pair": {
            "type": "boolean",
            "description": "Whether this photo has a stereo pair",
            "nullable": true
          },
          "stereo_pair_id": {
            "type": "integer",
            "description": "Stereo pair photo ID if available",
            "format": "int32",
            "nullable": true
          },
          "lighting_conditions": {
            "type": "string",
            "description": "Lighting conditions (e.g., \"golden_hour\", \"midday\", \"evening\")",
            "nullable": true
          },
          "location_visits": {
            "type": "integer",
            "description": "Number of times rover visited this location",
            "format": "int32",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "Photo-specific computed metadata"
      },
      "PhotoRelationships": {
        "type": "object",
        "properties": {
          "rover": {
            "$ref": "#/components/schemas/ResourceReference"
          },
          "camera": {
            "$ref": "#/components/schemas/CameraReference"
          }
        },
        "additionalProperties": false,
        "description": "Photo relationships (related resources)"
      },
      "PhotoResource": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "Unique photo identifier",
            "format": "int32"
          },
          "type": {
            "type": "string",
            "description": "Resource type (always \"photo\")",
            "nullable": true
          },
          "attributes": {
            "$ref": "#/components/schemas/PhotoAttributes"
          },
          "relationships": {
            "$ref": "#/components/schemas/PhotoRelationships"
          },
          "meta": {
            "$ref": "#/components/schemas/PhotoMeta"
          }
        },
        "additionalProperties": false,
        "description": "Photo resource in v2 API format with clear separation of attributes and relationships"
      },
...
      "PhotoTelemetry": {
        "type": "object",
        "properties": {
          "mast_azimuth": {
            "type": "number",
            "description": "Mast azimuth angle (horizontal rotation in degrees)",
            "format": "float",
            "nullable": true
          },
          "mast_elevation": {
            "type": "number",
            "description": "Mast elevation angle (vertical tilt in degrees)",
            "format": "float",
            "nullable": true
          },
          "spacecraft_clock": {
            "type": "number",
            "description": "Spacecraft clock at time of capture",
            "format": "float",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "Camera telemetry data"
      },
...
      "ResourceReference": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Resource identifier (slug for rovers, id for others)",
            "nullable": true
          },
          "type": {
            "type": "string",
            "description": "Resource type",
            "nullable": true
          },
          "attributes": {
            "description": "Attributes of the related resource (when included)",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "Reference to a related resource"
      },
      "ResponseLinks": {
        "type": "object",
        "properties": {
          "self": {
            "type": "string",
            "description": "Link to the current resource/page",
            "nullable": true
          },
          "next": {
            "type": "string",
            "description": "Link to the next page",
            "nullable": true
          },
          "previous": {
            "type": "string",
            "description": "Link to the previous page",
            "nullable": true
          },
          "first": {
            "type": "string",
            "description": "Link to the first page",
            "nullable": true
          },
          "last": {
            "type": "string",
            "description": "Link to the last page",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "Related resource links for navigation"
      },
      "ResponseMeta": {
        "type": "object",
        "properties": {
          "total_count": {
            "type": "integer",
            "description": "Total count of resources matching the query (before pagination)",
            "format": "int32",
            "nullable": true
          },
          "returned_count": {
            "type": "integer",
            "description": "Number of resources returned in this response",
            "format": "int32"
          },
          "query": {
            "type": "object",
            "additionalProperties": {},
            "description": "The query parameters that were applied",
            "nullable": true
          },
          "timestamp": {
            "type": "string",
            "description": "Response timestamp",
            "format": "date-time"
          }
        },
        "additionalProperties": false,
        "description": "Metadata about the response"
      },
...
      "ValidationError": {
        "type": "object",
        "properties": {
          "field": {
            "type": "string",
            "description": "The field/parameter name that failed validation",
            "nullable": true
          },
          "value": {
            "description": "The invalid value that was provided",
            "nullable": true
          },
          "message": {
            "type": "string",
            "description": "A human-readable error message",
            "nullable": true
          },
          "example": {
            "type": "string",
            "description": "An example of a valid value",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "Field-level validation error details"
      }
    },
    "securitySchemes": {
      "ApiKey": {
        "type": "apiKey",
        "description": "API key authentication. Get your key from the dashboard after signing in.",
        "name": "X-API-Key",
        "in": "header"
      }
    }
  },
  "security": [
    {
      "ApiKey": []
    }
  ],
  "servers": [
    {
      "url": "https://api.marsvista.dev",
      "description": "Production API"
    }
  ]
}